Refactored the clock and location arg test
