+++ src/tests/reachability/location_validation.rs
@@ -1,12 +1,12 @@
 #[cfg(test)]
 mod reachability_parser_location_validation {
     use crate::{
-        extract_system_rep::get_system_recipe,
-        tests::reachability::helper_functions::reachability_test_helper_functions, xml_parser,
-        JsonProjectLoader, ModelObjects::representations::QueryExpression, System,
-        XmlProjectLoader,
+        tests::reachability::helper_functions::reachability_test_helper_functions::{
+            self, create_system_recipe_and_machine,
+        },
+        ModelObjects::representations::QueryExpression,
+        System,
     };
-    use edbm::util::constraints::ClockIndex;
     use test_case::test_case;
 
     // These tests check that the parser only accepts location arguments with existing locations.
@@ -18,23 +18,10 @@ mod reachability_parser_location_validation {
     #[test_case("NOTCORRECTNAME";
     "The location NOTCORRECTNAME in the state does not exist in the model")]
     fn query_parser_checks_invalid_locations(location_str: &str) {
-        let folder_path = "samples/json/EcdarUniversity".to_string();
-        let mut comp_loader = if xml_parser::is_xml_project(&folder_path) {
-            XmlProjectLoader::new(folder_path)
-        } else {
-            JsonProjectLoader::new(folder_path)
-        }
-        .to_comp_loader();
         let mock_model = Box::new(QueryExpression::VarName("Adm2".to_string()));
-        let mut dim: ClockIndex = 0;
-        let mut quotient_index = None;
-        let machine = get_system_recipe(
-            &mock_model,
-            &mut (*comp_loader),
-            &mut dim,
-            &mut quotient_index,
-        );
-        let system = machine.clone().compile(dim).unwrap();
+        let folder_path = "samples/json/EcdarUniversity";
+        let (machine, system) = create_system_recipe_and_machine(*mock_model, folder_path);
+
         let mock_state = Box::new(QueryExpression::State(
             reachability_test_helper_functions::string_to_locations(location_str),
             None,
@@ -49,24 +36,11 @@ mod reachability_parser_location_validation {
     "The location L20 in the state exists in the model")]
     #[test_case("L23";
     "The location L23 in the state exists in the model")]
-    fn query_parser_checks_valid_clock_variables(location_str: &str) {
-        let folder_path = "samples/json/EcdarUniversity".to_string();
-        let mut comp_loader = if xml_parser::is_xml_project(&folder_path) {
-            XmlProjectLoader::new(folder_path)
-        } else {
-            JsonProjectLoader::new(folder_path)
-        }
-        .to_comp_loader();
+    fn query_parser_checks_valid_locations(location_str: &str) {
         let mock_model = Box::new(QueryExpression::VarName("Adm2".to_string()));
-        let mut dim: ClockIndex = 0;
-        let mut quotient_index = None;
-        let machine = get_system_recipe(
-            &mock_model,
-            &mut (*comp_loader),
-            &mut dim,
-            &mut quotient_index,
-        );
-        let system = machine.clone().compile(dim).unwrap();
+        let folder_path = "samples/json/EcdarUniversity";
+        let (machine, system) = create_system_recipe_and_machine(*mock_model, folder_path);
+
         let mock_state = Box::new(QueryExpression::State(
             reachability_test_helper_functions::string_to_locations(location_str),
             None,
